+++ src/hooks/useGames.ts
@@ -2,12 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import type { GameQuery } from "../App";
 import type { FetchResponse } from "../services/api-client";
 import apiClient from "../services/api-client";
-
-export interface Platform {
-  id: number;
-  name: string;
-  slug: string;
-}
+import type { Platform } from "./usePlatforms";
 
 export interface Game {
   id: number;
