+++ src/hooks/useGenres.ts
@@ -19,3 +19,4 @@ const useGenres = () =>
   });
 
 export default useGenres;
+
