+++ src/hooks/usePlatforms.ts
@@ -3,7 +3,7 @@ import apiClient from "../services/api-client";
 import type { FetchResponse } from "../services/api-client";
 import platforms from "../data/platforms";
 
-interface Platform {
+export interface Platform {
   id: number;
   name: string;
   slug: string;
